+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Control mapping using leafwing-input-manager
+
+### Changed
+
+- Move ship spawning and system registration from main to plugin
+
 ## [0.0.6] - 2023-09-29
 
 ### Changed